search bar people fixed
